requirements.txt
nextcord==2.6.0
google-api-python-client==2.128.0
python-dotenv==1.0.1
google-auth>=2.0.0
typing-extensions>=4.0.0
aiohttp>=3.8.0